// src/app/pages/patient/model/patient.model.ts
export class PatientModel {
    id: string;
    dateOfBirth: Date;
    address: string;
    bloodType: string;
    height: string;
    bloodPressure: string;
    email: string;
    fullName: string;
    phonenumber: string;
    password: string;
    gender: string;
    userId: string;
    role: string;

    constructor() {}
}